requirements/common-requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==0.16.1
python-dotenv==0.11.0
requests==2.22.0
